refactor(socket): socket code moved to controllers
